refactor(ui): 重构数据库节点显示逻辑
- 修改 DBNode 类，增加 dbSize 字段并调整构造函数
- 更新 JedisClusterClient 和 JedisPoolClient 中的 dbSize 方法返回类型
-调整 ReportController 和 ServerTabController 中的相关代码以适应新的 dbSize 返回类型
- 修改 Util 类中的 keyspaceParseDb 方法，返回值类型改为 Long

+++ rdm-ui/src/main/java/xyz/hashdog/rdm/ui/controller/ReportController.java
@@ -453,7 +453,7 @@ public class ReportController extends BaseKeyController<ServerTabController> imp
             List<InfoTable> infos= Util.parseInfoOutput(infoStr);
             this.list=infos;
             Map<String, String> map = infos.stream().filter(e->Constant.REDIS_INFO_KEYS.contains(e.getKey())).collect(Collectors.toMap(InfoTable::getKey, InfoTable::getValue));
-            List<Tuple2<Integer,Integer>> dbSizeList = new ArrayList<>();
+            List<Tuple2<Integer,Long>> dbSizeList = new ArrayList<>();
             infos.stream().filter(e->Constant.INFO_KEYSPACE.equals(e.getType())).forEach(e->dbSizeList.add(Util.keyspaceParseDb(e.getKey(),e.getValue())));
             Platform.runLater(()-> {
                 //bar数据更新
@@ -470,10 +470,10 @@ public class ReportController extends BaseKeyController<ServerTabController> imp
                 Tuple2<Double, String> barMemoryTu = Util.convertMemorySize(map.get(Constant.REDIS_INFO_USED_MEMORY));
                 barMemory.setText(String.format("%s%s",Util.format(barMemoryTu.t1(),2),barMemoryTu.t2()));
                 barMemory.setTooltip(GuiUtil.textTooltip(String.format(language("server.report.bar.memory")+" %s%s",Util.format(barMemoryTu.t1(),4),barMemoryTu.t2())));
-                int keyTotalSize = dbSizeList.stream().mapToInt(Tuple2::t2).sum();
+                long keyTotalSize = dbSizeList.stream().mapToLong(Tuple2::t2).sum();
                 barKey.setText(String.valueOf(keyTotalSize));
                 StringBuilder barKeyTooltip = new StringBuilder(String.format(language("server.report.bar.key")+" %s", keyTotalSize));
-                for (Tuple2<Integer, Integer> tus : dbSizeList) {
+                for (Tuple2<Integer, Long> tus : dbSizeList) {
                     barKeyTooltip.append(System.lineSeparator());
                     barKeyTooltip.append(String.format("DB%s: %s", tus.t1(), tus.t2()));
                 }
@@ -502,7 +502,7 @@ public class ReportController extends BaseKeyController<ServerTabController> imp
     }
 
 
-    private void updateLineCharData(double memoryValue,int keyTotalSize) {
+    private void updateLineCharData(double memoryValue,long keyTotalSize) {
         // 获取当前时间
         String timeLabel = LocalTime.now().format(DateTimeFormatter.ofPattern("HH:mm:ss"));
         // 添加新数据点
